Move some local data source queries to the correct class
