+++ core/src/main/java/org/hisp/dhis/android/sdk/synchronization/data/trackedentityinstance/TrackedEntityInstanceLocalDataSource.java
@@ -3,10 +3,8 @@ package org.hisp.dhis.android.sdk.synchronization.data.trackedentityinstance;
 
 import com.raizlabs.android.dbflow.sql.builder.Condition;
 import com.raizlabs.android.dbflow.sql.language.Select;
-
-import org.hisp.dhis.android.sdk.persistence.models.Enrollment;
-import org.hisp.dhis.android.sdk.persistence.models.Enrollment$Table;
 import org.hisp.dhis.android.sdk.persistence.models.TrackedEntityInstance;
+import org.hisp.dhis.android.sdk.persistence.models.TrackedEntityInstance$Table;
 
 import java.util.List;
 
@@ -15,8 +13,13 @@ public class TrackedEntityInstanceLocalDataSource {
         trackedEntityInstance.save();
     }
 
-    public List<Enrollment> getEnrollments(long localId) {
-        return new Select().from(Enrollment.class).where(Condition.column(Enrollment$Table.LOCALTRACKEDENTITYINSTANCEID).
-                is(localId)).queryList();
+    public List<TrackedEntityInstance> getAllLocalTeis() {
+        return new Select().from(TrackedEntityInstance.class).where(Condition.column(TrackedEntityInstance$Table.FROMSERVER).
+                is(false)).queryList();
+    }
+
+    public TrackedEntityInstance getTrackedEntityInstance(String trackedEntityInstanceUid) {
+        return new Select().from(TrackedEntityInstance.class).where(Condition.column
+                (TrackedEntityInstance$Table.TRACKEDENTITYINSTANCE).is(trackedEntityInstanceUid)).querySingle();
     }
 }
