+++ core/src/main/java/org/hisp/dhis/android/sdk/synchronization/domain/enrollment/IEnrollmentRepository.java
@@ -15,5 +15,5 @@ public interface IEnrollmentRepository {
 
     Enrollment getEnrollment(String enrollmentUid);
 
-    TrackedEntityInstance getTrackedEntityInstance(String trackedEntityInstanceUid);
+    List<Enrollment> getEnrollments(long trackedEntityInstanceLocalId);
 }
